rkt: Remove dead code
Some of the contents of images.go were used by the image subcommands,
so move the code to the separate file.
